Add code to format CREATE VIEW | RULE statement
After upgrade, the format in which pg_dumpall dumps the
CREATE VIEW and CREATE RULE statement has changed. This commit format
the dump to match the old format.

Old format is of the below format:
CREATE VIEW view_name AS
SELECT ....; (in single line)

CREATE RULE rule_name AS ....;(in single and same line as CREATE)

New format is of the below format:

CREATE VIEW view_name AS
SELECT
    ...
    ...
    ...;

CREATE RULE rule_name AS
    ...
    ...;